cleanup: Move get_exit_code to libs and remove throwaway code
